Generate an error if offset is too accurate in renderRfc3339
RFC3339 only supports offsets as accurate as a minute, but Local
provides the offset in seconds. This means that the rendering
of RFC3339 strings in utc is lossy, i. e. a value rendered and could be
different when parsed, as the RFC3339 string wouldn't be able to
represent it accurate enough. To prevent this we now refuse to render
Local values that would result in data loss by throwing an UtcException.